+++ folio-spring-system-user/src/test/java/org/folio/spring/systemuser/SystemUserServiceTest.java
@@ -63,8 +63,10 @@ class SystemUserServiceTest {
   private static final Instant REFRESH_TOKEN_EXPIRATION = NOW.plus(7, ChronoUnit.DAYS);
   private static final Instant CUSTOM_TOKEN_EXPIRATION = TOKEN_EXPIRATION.minus(12, ChronoUnit.HOURS);
   private static final String MOCK_TOKEN = "test_token";
-  private static final String CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID
-    = "Cannot retrieve okapi token for tenant: tenantId";
+  private static final String CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME
+    = "Cannot retrieve okapi token for system user: username";
+  private static final String CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU
+    = "Cannot retrieve okapi token for system user: username [tenantId=diku]";
   private final ResponseEntity<AuthnClient.LoginResponse> expectedResponse =
     Mockito.spy(ResponseEntity.of(Optional.of(
       new AuthnClient.LoginResponse(TOKEN_EXPIRATION.toString(), REFRESH_TOKEN_EXPIRATION.toString()))));
@@ -186,9 +188,9 @@ class SystemUserServiceTest {
 
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService
-      .authSystemUser("tenantId", "username", "password"))
+      .authSystemUser("diku", "username", "password"))
       .isInstanceOf(SystemUserAuthorizationException.class)
-      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID);
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU);
   }
 
   @Test
@@ -215,9 +217,9 @@ class SystemUserServiceTest {
 
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService
-      .authSystemUser("tenantId", "username", "password"))
+      .authSystemUser("diku", "username", "password"))
       .isInstanceOf(SystemUserAuthorizationException.class)
-      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID);
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU);
   }
 
   @Test
@@ -240,9 +242,9 @@ class SystemUserServiceTest {
 
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService
-      .authSystemUser("tenantId", "username", "password"))
+      .authSystemUser("diku", "username", "password"))
       .isInstanceOf(SystemUserAuthorizationException.class)
-      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID);
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU);
   }
 
   @Test
@@ -281,7 +283,7 @@ class SystemUserServiceTest {
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService.authSystemUser(systemUser)).isInstanceOf(
         SystemUserAuthorizationException.class)
-      .hasMessage("Cannot retrieve okapi token for tenant: username");
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME);
   }
 
   @Test
@@ -292,16 +294,16 @@ class SystemUserServiceTest {
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService.authSystemUser(systemUser)).isInstanceOf(
         SystemUserAuthorizationException.class)
-      .hasMessage("Cannot retrieve okapi token for tenant: username");
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME);
   }
 
   @Test
   void overloaded_authSystemUser_when_loginExpiry_Returns400Response() {
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService
-      .authSystemUser("tenantId", "username", "password"))
+      .authSystemUser("diku", "username", "password"))
       .isInstanceOf(SystemUserAuthorizationException.class)
-      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID);
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU);
   }
 
   @Test
@@ -309,9 +311,9 @@ class SystemUserServiceTest {
     when(contextBuilder.forSystemUser(any())).thenReturn(context);
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService
-      .authSystemUser("tenantId", "username", "password"))
+      .authSystemUser("diku", "username", "password"))
       .isInstanceOf(SystemUserAuthorizationException.class)
-      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID);
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU);
   }
 
   @Test
@@ -319,9 +321,9 @@ class SystemUserServiceTest {
     when(contextBuilder.forSystemUser(any())).thenReturn(context);
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService
-      .authSystemUser("tenantId", "username", "password"))
+      .authSystemUser("diku", "username", "password"))
       .isInstanceOf(SystemUserAuthorizationException.class)
-      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID);
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU);
   }
 
   @Test
@@ -334,16 +336,16 @@ class SystemUserServiceTest {
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService.authSystemUser(systemUser)).isInstanceOf(
         SystemUserAuthorizationException.class)
-      .hasMessage("Cannot retrieve okapi token for tenant: username");
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME);
   }
 
   @Test
   void overloaded_authSystemUser_when_loginExpiry_and_tokenLegacy_both_notFound() {
     var systemUserService = systemUserService(systemUserProperties());
     assertThatThrownBy(() -> systemUserService
-      .authSystemUser("tenantId", "username", "password"))
+      .authSystemUser("diku", "username", "password"))
       .isInstanceOf(SystemUserAuthorizationException.class)
-      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_TENANT_TENANT_ID);
+      .hasMessage(CANNOT_RETRIEVE_OKAPI_TOKEN_FOR_USERNAME_AND_DIKU);
   }
 
   private static SystemUser systemUserValue() {